% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crawl_sentence.R
\name{crawl_sentence}
\alias{crawl_sentence}
\title{Take a sentence and follow the dependencies}
\usage{
crawl_sentence(s)
}
\arguments{
\item{s}{a data.frame containing the results of one (1) parsed spacy sentence}
}
\value{
list with original parsed sentence + added dependency parsing
}
\description{
Take a sentence and follow the dependencies
}
