% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_empty_sentence_parse_list}
\alias{create_empty_sentence_parse_list}
\title{Create a list to iterate over a sentence and store items}
\usage{
create_empty_sentence_parse_list(N)
}
\arguments{
\item{N}{integer value that sets vector length}
}
\value{
list with named empty vectors
}
\description{
Takes an integer value equivalent to token length of sentence and generates named list
}
