% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_entity_extract_V2.R
\name{custom_entity_extract2}
\alias{custom_entity_extract2}
\title{Takes a parsed spacy document and performs dependency parsing}
\usage{
custom_entity_extract2(
  x,
  concatenator = "_",
  file = NULL,
  cl = 1,
  keep_entities = c("ORG", "GPE", "PERSON"),
  return_to_memory = T
)
}
\arguments{
\item{x}{parsed spacy document}

\item{concatenator}{how entity parts are concatenated (defaults to "_")}

\item{file}{location where an list object with an edgelist and nodelist should be saved as .RDS file}

\item{cl}{number of cores to crawl sentences in parallel (defaults to 1)}

\item{keep_entities}{character vector of spacy entity types to retain, defaults to people (PERSON), organizations (ORG), and geographic entities (GPE)}

\item{return_to_memory}{boolean for whether function should return final result as workspace object}
}
\value{
data frame with original parsed sentence + added dependency parsing
}
\description{
Takes a parsed spacy document and performs dependency parsing
}
