% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_dictionary.R
\name{download_dictionary}
\alias{download_dictionary}
\title{Downloads the SCOWL dictionary, which has a larger vocabulary than Qdap}
\usage{
download_dictionary()
}
\description{
Downloads the SCOWL dictionary, which has a larger vocabulary than Qdap
}
