% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_10p}
\alias{sample_10p}
\title{Sample of 10 pages from a groundwater plan
Parsed by spacy (spacy_en_web_lg) and stored as data.frame}
\format{
An object of class \code{"spacyr_parsed"}.
}
\usage{
data(sample_10p)
}
\description{
Sample of 10 pages from a groundwater plan
Parsed by spacy (spacy_en_web_lg) and stored as data.frame
}
\examples{
data(sample_10p)
tokens <- sample_10p$token
}
\keyword{datasets}
