% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbnet_port.R
\name{verbnet_port}
\alias{verbnet_port}
\title{Links to VerbNet 3.3 and uses the classification structure to create a data.table
where each row is a verb, and the columns represent the types and classifications in which that
verb can be found. Also adds a designation for "helping verbs."}
\usage{
verbnet_port(zipdestfile, folder_dest)
}
\arguments{
\item{zipdestfile}{A filepath on which to store the zipfiles downloaded from VerbNet}

\item{folder_dest}{A filepath for the folder in which to put the extracted VerbNet files}
}
\value{
Returns the data.table of verbs and their classifications.
}
\description{
Links to VerbNet 3.3 and uses the classification structure to create a data.table
where each row is a verb, and the columns represent the types and classifications in which that
verb can be found. Also adds a designation for "helping verbs."
}
