% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_entities.R
\name{clean_entities}
\alias{clean_entities}
\title{Clean entity name strings}
\usage{
clean_entities(
  v,
  remove_nums = TRUE,
  remove_trailing_s = TRUE,
  concatenator = "_"
)
}
\arguments{
\item{v}{a vector of entity names}

\item{remove_nums}{A boolean. If TRUE, sets entities that contain no letters to an empty string. If FALSE, sets entities that contain no letters or numbers to an empty string.}

\item{remove_trailing_s}{A boolean. If TRUE, removes trailing instances of concatenator+s, or apostrophe+s in each element of v.}

\item{concatenator}{Defaults to an underscore. Use regex notation. The concatenator used in elements of v.}
}
\value{
a cleaned vector of entity names
}
\description{
Normalizes a character vector of entity names by converting mathematical
Unicode characters to ASCII letters, removing non-word characters,
collapsing consecutive concatenators, and stripping leading/trailing
concatenators.
}
