% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_entities.R
\name{clean_entities}
\alias{clean_entities}
\title{Take a vector and clean the strings using regex.
Uppercase or lowercase math font is converted to uppercase or lowercase letters, respectively.
Any trailing "'s" at the end of the entity name is removed.
All non-word characters are removed.
Consecutive underscores are collapsed to a single underscore.
Leading and trailing underscores are removed.
Entities that have no letters are removed, if remove_nums is set to T.}
\usage{
clean_entities(v, remove_nums = T)
}
\arguments{
\item{v}{a vector of entity names}

\item{remove_nums}{A boolean. If T, sets entities that contain no letters to an empty string. If F, sets entities that contain no letters or numbers to an empty string.}
}
\value{
a cleaned vector of entity names
}
\description{
Take a vector and clean the strings using regex.
Uppercase or lowercase math font is converted to uppercase or lowercase letters, respectively.
Any trailing "'s" at the end of the entity name is removed.
All non-word characters are removed.
Consecutive underscores are collapsed to a single underscore.
Leading and trailing underscores are removed.
Entities that have no letters are removed, if remove_nums is set to T.
}
