% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_networks.R
\name{combine_networks}
\alias{combine_networks}
\title{Superimposes multiple textNet igraphs into a single igraph.}
\usage{
combine_networks(textnet_igraphs, mode)
}
\arguments{
\item{textnet_igraphs}{List of textnet igraphs, the outputs of export_to_network(export_format = "igraph)}

\item{mode}{Either "multiplex" or "weighted" depending on the desired exported network format. If "weighted", collapses multiplex edges into a single weighted edge}
}
\value{
list of all entities and lemmas in the corpus, along with their average normalized prevalence as a fraction of a plan
}
\description{
Superimposes multiple textNet igraphs into a single igraph.
}
