% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eng_words-data.R
\docType{data}
\name{eng_words}
\alias{eng_words}
\title{English dictionary fom the SCOWL 2020.12.07 dictionary created by Kevin Atkinson
Files with extension <= 60, with filenames that include "variant", "american", "british", canadian", or "australian"}
\format{
A character vector
}
\source{
\href{https://sourceforge.net/projects/wordlist/files/SCOWL/2020.12.07/scowl-2020.12.07.zip/download}{SCOWL 2020.12.07}
}
\usage{
data(eng_words)
}
\description{
English dictionary fom the SCOWL 2020.12.07 dictionary created by Kevin Atkinson
Files with extension <= 60, with filenames that include "variant", "american", "british", canadian", or "australian"
}
\examples{
data(eng_words)
}
\references{
Kevin Atkinson 2020
}
\keyword{scowl}
