% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entity_consolidate_replicate.R
\name{entity_consolidate_replicate}
\alias{entity_consolidate_replicate}
\title{Modified version of spacyr::entity_consolidate.
spacyr::entity_consolidate returns a data.frame that no longer has head_token_id and dep_rel data.
Instead, this function basically breaks off a feature of the original function and adds a new column to the original spacyr data.frame that is the concatenated entity.
This feature is inefficient in that the concatenated entity is then replicated multiple times, but this does seem to be the easiest way to preserve the other data.}
\usage{
entity_consolidate_replicate(x, concatenator = "_", remove = NULL)
}
\arguments{
\item{x}{parsed spacy document}

\item{concatenator}{A character that separates string segments when they are collapsed into a single entity. Defaults to "_"}

\item{remove}{regex formatted strings to remove as entity components (like "the" in "the Seattle Supersonics")}
}
\value{
original data frame with added column for concatenated entity
}
\description{
Modified version of spacyr::entity_consolidate.
spacyr::entity_consolidate returns a data.frame that no longer has head_token_id and dep_rel data.
Instead, this function basically breaks off a feature of the original function and adds a new column to the original spacyr data.frame that is the concatenated entity.
This feature is inefficient in that the concatenated entity is then replicated multiple times, but this does seem to be the easiest way to preserve the other data.
}
