% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_acronyms.R
\name{find_acronyms}
\alias{find_acronyms}
\title{Take a character vector and detect all parenthetical statements in which an acronym is defined within the character vector.}
\usage{
find_acronyms(str)
}
\arguments{
\item{str}{A character vector}
}
\value{
a data table with a "name" column and an "acronym" column representing its associated acronym.
Each row corresponds to a unique match in the document.
}
\description{
Take a character vector and detect all parenthetical statements in which an acronym is defined within the character vector.
}
