% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_acronyms.R
\name{find_acronyms}
\alias{find_acronyms}
\title{Take a character vector and detect all parenthetical statements in which an acronym is uniquely defined within the character vector.}
\usage{
find_acronyms(str, table_text = NULL)
}
\arguments{
\item{str}{A character vector}

\item{table_text}{Optional; used to detect acronyms from tables. A character vector representing the contents of a unique text file, such as the output of pdftools::pdf_text. Each element of the vector corresponds to a page of a pdf. It is recommended to use raw text output, rather than the results of textNet::pdf_clean, as the cleaning process removes white spaces that could hinder the detection of acronym tables. Single-letter acronyms are excluded. Parenthetically defined acronyms take precedence over table-defined acronyms.}
}
\value{
a data table with a "name" column and an "acronym" column representing its associated acronym.
Each row corresponds to a unique match in the document.
}
\description{
Take a character vector and detect all parenthetical statements in which an acronym is uniquely defined within the character vector.
}
