% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_new_parsed-data.R
\docType{data}
\name{old_new_parsed}
\alias{old_new_parsed}
\title{Sample of parsed data extracted from two versions of a publicly available groundwater plan,
stored as a data frame. Equivalent data format to the output of parse_text}
\format{
An object of class \code{"spacyr_parsed"}.
}
\usage{
data(old_new_parsed)
}
\description{
Sample of parsed data extracted from two versions of a publicly available groundwater plan,
stored as a data frame. Equivalent data format to the output of parse_text
}
\examples{
data(old_new_parsed)
tokens <- old_new_parsed$token
}
\keyword{datasets}
