% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_new_text-data.R
\docType{data}
\name{old_new_text}
\alias{old_new_text}
\title{A sample of text data extracted from pdfs representing two versions of a publicly available groundwater plan,
stored as a named list. Each element in the list represents a different document,
and each string represents the text of one page. Equivalent data format to the output of pdf_clean}
\format{
An object of class \code{"list"}.
}
\usage{
data(old_new_text)
}
\description{
A sample of text data extracted from pdfs representing two versions of a publicly available groundwater plan,
stored as a named list. Each element in the list represents a different document,
and each string represents the text of one page. Equivalent data format to the output of pdf_clean
}
\examples{
data(old_new_text)
}
\keyword{datasets}
