% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_text.R
\name{parse_text}
\alias{parse_text}
\title{Creates an edgelist and nodelist for each document}
\usage{
parse_text(
  ret_path,
  keep_hyph_together = F,
  phrases_to_concatenate = NA,
  concatenator = "_",
  text_list,
  parsed_filenames,
  overwrite = T,
  custom_entities = NULL,
  model = "en_core_web_lg"
)
}
\arguments{
\item{ret_path}{filepath to use for Sys.setenv reticulate python call. Note: Python and miniconda must already be installed.}

\item{keep_hyph_together}{Set to true to replace hyphens within a single word with underscores. Defaults to false.}

\item{phrases_to_concatenate}{character vector of phrases, in which each element is a string consisting of tokens separated by spaces. These are replaced with their concatenated version in order, from left to right. It is suggested that the most specific phrases, with the most words, are arranged at the left.}

\item{concatenator}{This is a character or string that will be used to replace the spaces in the phrases_to_concatenate.}

\item{text_list}{This is a named list, an object of the type resulting from pdf_clean, in which each list element is a document, and each string within a list element represents the text on one page}

\item{parsed_filenames}{This is a character vector in which each element represents a filepath associated with its respective document.
The parsed data will be exported to these files.}

\item{overwrite}{A boolean. Whether to overwrite existing files}

\item{custom_entities}{A named list. This does not overwrite the entity determination of the NLP engine, but rather catches user-defined entities that are not otherwise detected by the engine. Best used in combination with phrases_to_concatenate, since the custom entity label will only be applied if the entire token matches the definition. Does not search multiple consecutive tokens to define a match. These will be applied to all documents.}
}
\value{
A data.frame of tokens. For more information on the format, see the spacyr::spacy_parse help file
}
\description{
Creates an edgelist and nodelist for each document
}
