% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_clean.R
\name{pdf_clean}
\alias{pdf_clean}
\title{Extract and clean text from PDF files}
\usage{
pdf_clean(
  pdfs,
  keep_pages = NULL,
  ocr = FALSE,
  maxchar = 10000,
  export_paths = NULL,
  return_to_memory = TRUE,
  suppressWarn = FALSE,
  auto_headfoot_remove = TRUE,
  sentence_shift = FALSE
)
}
\arguments{
\item{pdfs}{a vector of file names}

\item{keep_pages}{By default, NULL keeps all pages. Alternatively, the user can specify a list of logical or numeric vectors of pages to keep in the exported file.
When specified, the length of keep_pages must be equivalent to the length of pdfs.
Each logical vector in keep_pages must have a length equivalent to the number of pages in the corresponding input pdf.
All pages with a value of TRUE will be included in the exported file.
Each numeric vector in keep_pages enumerates specific page numbers to keep.
For numeric vectors in keep_pages, duplicate values will be returned only once, and the vector will be sorted before processing.}

\item{ocr}{A logical value: TRUE to run ocr image-to-text detection on pages with fewer than 20 characters;
FALSE to not run ocr on any pages.}

\item{maxchar}{A numeric value representing the maximum allowable number of characters per page,
with a default of 10,000. Pages on which the number of characters exceeds maxchar will have the output text
set to NA.}

\item{export_paths}{A vector of filenames to which to save the text outputs.
Defaults to NULL, which does not save the output to a file.}

\item{return_to_memory}{A logical value representing whether or not to return
the output text to memory. Defaults to TRUE.}

\item{suppressWarn}{Boolean; FALSE by default. Setting this to TRUE suppresses pdf_text conversion warnings, which can be verbose.}

\item{auto_headfoot_remove}{Boolean. Tries to remove headers and footers if TRUE. May remove small amounts of paragraph text, especially if there is no header on a page. Defaults to TRUE to avoid having entities that appear consistently in headers be artificially inflated in the textnet_extract network. Recommend keeping this parameter TRUE unless using a separate header/footer removal tool.}

\item{sentence_shift}{Boolean. If TRUE, tries to shift a beginning partial sentence to the previous page.}
}
\value{
If return_to_memory is TRUE, returns a list with a length equal
to the length of the pdfs parameter. Each element in the list is a character vector
with a length equal to the number of pages in its respective document. If export_paths
is not null, each character vector of text will be saved to a separate RDS file, where the
number of files is equal to the number of pdfs.
}
\description{
Wraps \code{\link[pdftools:pdf_text]{pdftools::pdf_text()}} to read one or more PDF files into R, with
optional header and footer removal, page filtering, OCR support, and
sentence-boundary shifting.
}
