% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_clean.R
\name{pdf_clean}
\alias{pdf_clean}
\title{Take a vector and clean the strings using regex.
Uppercase or lowercase math font is converted to uppercase or lowercase letters, respectively.
Any trailing "'s" at the end of the entity name is removed.
All non-word characters are removed.
Consecutive underscores are collapsed to a single underscore.
Leading and trailing underscores are removed.
Entities that have no letters are removed, if remove_nums is set to T.}
\usage{
pdf_clean(
  pdfs,
  keep_pages = NULL,
  ocr = F,
  maxchar = 10000,
  export_paths = NULL,
  return_to_memory = T,
  suppressWarn = F,
  auto_headfoot_remove = T
)
}
\arguments{
\item{pdfs}{a vector of file names}

\item{keep_pages}{By default, NULL keeps all pages. Alternatively, the user can specify a list of logical or numeric vectors of pages to keep in the exported file.
When specified, the length of keep_pages must be equivalent to the length of pdfs.
Each logical vector in keep_pages must have a length equivalent to the number of pages in the corresponding input pdf.
All pages with a value of T will be included in the exported file.
Each numeric vector in keep_pages enumerates specific page numbers to keep.
For numeric vectors in keep_pages, duplicate values will be returned only once, and the vector will be sorted before processing.}

\item{ocr}{A logical value: T to run ocr image-to-text detection on pages with fewer than 20 characters;
F to not run ocr on any pages.}

\item{maxchar}{A numeric value representing the maximum allowable number of characters per page,
with a default of 10,000. Pages on which the number of characters exceeds maxchar will have the output text
set to NA.}

\item{export_paths}{A vector of filenames to which to save the text outputs.
Defaults to NULL, which does not save the output to a file.}

\item{return_to_memory}{A logical value representing whether or not to return
the output text to memory. Defaults to T.}

\item{suppressWarn}{Boolean; F by default. Setting this to T suppresses pdf_text conversion warnings, which can be verbose.}

\item{auto_headfoot_remove}{Boolean. Tries to remove headers and footers if T. May remove small amounts of paragraph text, especially if there is no header on a page. Defaults to T to avoid having entities that appear consistently in headers be artificially inflated in the textnet_extract network. Recommend keeping this parameter T unless using a separate header/footer removal tool.}
}
\value{
If return_to_memory is T, returns a list with a length equal
to the length of the pdfs parameter. Each element in the list is a character vector
with a length equal to the number of pages in its respective document. If export_paths
is not null, each character vector of text will be saved to a separate RDS file, where the
number of files is equal to the number of pdfs.
}
\description{
Take a vector and clean the strings using regex.
Uppercase or lowercase math font is converted to uppercase or lowercase letters, respectively.
Any trailing "'s" at the end of the entity name is removed.
All non-word characters are removed.
Consecutive underscores are collapsed to a single underscore.
Leading and trailing underscores are removed.
Entities that have no letters are removed, if remove_nums is set to T.
}
