% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_20p-data.R
\docType{data}
\name{sample_20p}
\alias{sample_20p}
\title{Sample of 20 pages from a collection of groundwater plans
Parsed by spacy (spacy_en_web_lg) and stored as data.frame}
\format{
An object of class \code{"spacyr_parsed"}.
}
\usage{
data(sample_20p)
}
\description{
Sample of 20 pages from a collection of groundwater plans
Parsed by spacy (spacy_en_web_lg) and stored as data.frame
}
\examples{
data(sample_20p)
tokens <- sample_20p$token
}
\keyword{datasets}
