% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textnet_extract.R
\name{textnet_extract}
\alias{textnet_extract}
\title{Takes a parsed spacy document and uses dependency parsing to generate an edgelist, nodelist, verblist, and apposititve list}
\usage{
textnet_extract(
  x,
  concatenator = "_",
  file = NULL,
  cl = 1,
  keep_entities = c("ORG", "GPE", "PERSON"),
  return_to_memory = T,
  keep_incomplete_edges = F
)
}
\arguments{
\item{x}{parsed spacy document}

\item{concatenator}{how entity parts are concatenated (defaults to "_")}

\item{file}{location where an list object with an edgelist and nodelist should be saved as .RDS file}

\item{cl}{number of cores to crawl sentences in parallel (defaults to 1)}

\item{keep_entities}{character vector of spacy entity types to retain, defaults to people (PERSON), organizations (ORG), and geographic entities (GPE)}

\item{return_to_memory}{boolean for whether function should return final result as workspace object}

\item{keep_incomplete_edges}{Boolean. If T, keeps edges with only a source or target but not both}
}
\value{
data frame with original parsed sentence + added dependency parsing
}
\description{
Takes a parsed spacy document and uses dependency parsing to generate an edgelist, nodelist, verblist, and apposititve list
}
