% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_features.R
\name{top_features}
\alias{top_features}
\title{Finds the highest degree entities and most common lemmas in the network}
\usage{
top_features(files, from_file = FALSE)
}
\arguments{
\item{files}{vector of filepaths to igraph objects or list of igraph objects}

\item{from_file}{boolean whether files represent filepaths (TRUE) or igraph objects (FALSE)}
}
\value{
list of all entities and lemmas in the corpus, along with their average normalized prevalence as a fraction of a plan. For entities, this is the entity degree over the sum of all entity degrees in the plan, averaged across all plans
}
\description{
Finds the highest degree entities and most common lemmas in the network
}
