% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_features.R
\name{top_features}
\alias{top_features}
\title{Finds the most common entities and lemmas in the corpus}
\usage{
top_features(files, from_file = F)
}
\arguments{
\item{files}{vector of filepaths to igraph objects or list of igraph objects}

\item{from_file}{boolean whether files represent filepaths (T) or igraph objects (F)}
}
\value{
list of all entities and lemmas in the corpus, along with their average normalized prevalence as a fraction of a plan
}
\description{
Finds the most common entities and lemmas in the corpus
}
