% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verb_classifications-data.R
\docType{data}
\name{verb_classifications}
\alias{verb_classifications}
\title{Data downloaded from VerbNet 3.3 and organized by type and class, as in the script "verbnet_port.R".}
\format{
An object of class "data.table"
}
\source{
\href{https://verbs.colorado.edu/verb-index/vn3.3/}{VerbNet 3.3}
}
\usage{
data(verb_classifications)
}
\description{
Attributes from this table are bound to the verblist exported by textnet_extract.
}
\examples{
data(verb_classifications)
}
\references{
Schuler, Karin Kipper. 2005. VerbNet: A Broad-Coverage, Comprehensive Verb Lexicon. University of Pennsylvania.
}
\keyword{verbnet}
